refactor: 게임 시작 클래스를 따로 추가
