+++ src/main/java/racingcar/domain/Racing.java
@@ -18,14 +18,6 @@ public class Racing {
         this.random = new Random(System.currentTimeMillis());
     }
 
-    public static void main(String[] args) {
-        int carCount = InputView.nextInt("자동차 대수는 몇 대 인가요?");
-        int tryCount = InputView.nextInt("시도할 회수는 몇 회 인가요?");
-
-        Racing racing = new Racing(carCount, tryCount);
-        racing.start();
-    }
-
     public List<Car> makeCar(int carCount) {
         List<Car> carList = new ArrayList<>();
         for (int i = 0; i < carCount; i++) {
